api/views.py
from rest_framework import viewsets, status, permissions
from rest_framework.decorators import action
from rest_framework.response import Response
from rest_framework.authtoken.models import Token
from django.contrib.auth.models import User
from django.db.models import Q, Count
from django.utils import timezone
from accounts.models import UserProfile
from companies.models import Company
from jobs.models import Job
from applications.models import Application
from .serializers import (
    UserSerializer, UserProfileSerializer, CompanySerializer, 
    JobSerializer, JobCreateSerializer, ApplicationSerializer, 
    ApplicationCreateSerializer, ApplicationStatusUpdateSerializer,
    DashboardStatsSerializer
)