requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==11.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.32.3
